COM-2013 - Use card store for questionnaire answers
