+++ src/store/activities/reducers.ts
@@ -3,9 +3,6 @@ import {
   SET_ACTIVITY,
   SET_EXIT_CONFIRMATION_MODAL,
   RESET_ACTIVITY_REDUCER,
-  ADD_QUESTIONNAIRE_ANSWER,
-  REMOVE_QUESTIONNAIRE_ANSWER,
-  SET_QUESTIONNAIRE_ANSWERS_LIST,
   INC_GOOD_ANSWERS_COUNT,
   SET_ACTIVITY_HISTORIES,
   ActivityActionType,
@@ -15,29 +12,8 @@ import {
 const initialState: ActivityStateType = {
   activity: null,
   exitConfirmationModal: false,
-  questionnaireAnswersList: [],
   score: 0,
   activityHistories: [],
-};
-
-const applyAddQuestionnaireAnswer = (state, action) => {
-  const questionnaireAnswer = action.payload;
-  const indexOfQuestionnaireAnswer = state.questionnaireAnswersList.findIndex((qa =>
-    qa.card === questionnaireAnswer.card));
-
-  if (indexOfQuestionnaireAnswer !== -1) {
-    const newQuestionnaireAnswersList = [...state.questionnaireAnswersList];
-    newQuestionnaireAnswersList[indexOfQuestionnaireAnswer] = questionnaireAnswer;
-
-    return { ...state, questionnaireAnswersList: newQuestionnaireAnswersList };
-  }
-
-  return { ...state, questionnaireAnswersList: [...state.questionnaireAnswersList, questionnaireAnswer] };
-};
-
-const applyRemoveQuestionnaireAnswer = (state, action) => {
-  const card = action.payload;
-  return { ...state, questionnaireAnswersList: state.questionnaireAnswersList.filter(qa => qa.card !== card) };
 };
 
 export const activities = (
@@ -49,16 +25,10 @@ export const activities = (
       return { ...state, activity: action.payload };
     case SET_EXIT_CONFIRMATION_MODAL:
       return { ...state, exitConfirmationModal: action.payload };
-    case ADD_QUESTIONNAIRE_ANSWER:
-      return applyAddQuestionnaireAnswer(state, action);
-    case REMOVE_QUESTIONNAIRE_ANSWER:
-      return applyRemoveQuestionnaireAnswer(state, action);
     case SET_ACTIVITY_HISTORIES:
       return { ...state, activityHistories: action.payload };
     case RESET_ACTIVITY_REDUCER:
       return initialState;
-    case SET_QUESTIONNAIRE_ANSWERS_LIST:
-      return { ...state, questionnaireAnswersList: action.payload };
     case INC_GOOD_ANSWERS_COUNT:
       return { ...state, score: state.score + 1 };
     default:
