+++ src/store/activities/selectors.ts
@@ -1,13 +1,6 @@
-import { OPEN_QUESTION, QUESTION_ANSWER, SURVEY, TRANSITION } from '../../core/data/constants';
-import { QuestionnaireAnswerType } from '../../types/store/ActivityStoreType';
+import { TRANSITION } from '../../core/data/constants';
 
 const getCard = state => state.cards.cards[state.cards.cardIndex];
-
-const getQuestionnaireAnswer = (state): QuestionnaireAnswerType | null => {
-  const card = getCard(state);
-  if (!card || (![SURVEY, OPEN_QUESTION, QUESTION_ANSWER].includes(card.template))) return null;
-  return state.activities.questionnaireAnswersList.find(qa => qa.card === card._id) || null;
-};
 
 const getMaxProgress = state => state.cards.cards.filter(card => card.template !== TRANSITION).length;
 
@@ -22,7 +15,6 @@ const displayProgressBar = state => !!getCard(state) && getCard(state).template 
 
 export default {
   getCard,
-  getQuestionnaireAnswer,
   getMaxProgress,
   getProgress,
   displayProgressBar,
