+++ src/types/store/ActivityStoreType.ts
@@ -4,10 +4,7 @@ import { ActivityHistoryType } from '../ActivityHistoryType';
 // Actions types
 export const SET_ACTIVITY = 'SET_ACTIVITY';
 export const SET_EXIT_CONFIRMATION_MODAL = 'SET_EXIT_CONFIRMATION_MODAL';
-export const ADD_QUESTIONNAIRE_ANSWER = 'ADD_QUESTIONNAIRE_ANSWER';
-export const REMOVE_QUESTIONNAIRE_ANSWER = 'REMOVE_QUESTIONNAIRE_ANSWER';
 export const RESET_ACTIVITY_REDUCER = 'RESET_ACTIVITY_REDUCER';
-export const SET_QUESTIONNAIRE_ANSWERS_LIST = 'SET_QUESTIONNAIRE_ANSWERS_LIST';
 export const INC_GOOD_ANSWERS_COUNT = 'INC_GOOD_ANSWERS_COUNT';
 export const SET_ACTIVITY_HISTORIES = 'SET_ACTIVITY_HISTORIES';
 
@@ -18,21 +15,6 @@ export interface SetActivityType {
 export interface SetExitConfirmationModalType {
   type: typeof SET_EXIT_CONFIRMATION_MODAL,
   payload: boolean,
-}
-
-export interface AddQuestionnaireAnswerType {
-  type: typeof ADD_QUESTIONNAIRE_ANSWER,
-  payload: QuestionnaireAnswerType,
-}
-
-export interface RemoveQuestionnaireAnswerType {
-  type: typeof REMOVE_QUESTIONNAIRE_ANSWER,
-  payload: string,
-}
-
-export interface SetQuestionnaireAnswersListType {
-  type: typeof SET_QUESTIONNAIRE_ANSWERS_LIST,
-  payload: Array<QuestionnaireAnswerType>,
 }
 
 export interface SetActivityHistories {
@@ -51,23 +33,13 @@ export interface IncGoodAnswersCountType {
 export type ActivityActionType =
 SetActivityType |
 SetExitConfirmationModalType |
-AddQuestionnaireAnswerType |
-RemoveQuestionnaireAnswerType |
-SetQuestionnaireAnswersListType |
 SetActivityHistories;
 
 export type ActivityActionWithoutPayloadType = ResetActivityReducer | IncGoodAnswersCountType;
 
-export interface QuestionnaireAnswerType {
-  _id?: string,
-  card: string,
-  answerList: Array<string>,
-}
-
 export interface ActivityStateType {
   activity: ActivityType | null,
   exitConfirmationModal: boolean,
-  questionnaireAnswersList: Array<QuestionnaireAnswerType>,
   score: number
   activityHistories: Array<ActivityHistoryType>
 }
